requirements.txt
deep-translator
spacy
gTTS
google-cloud-texttospeech
azure-cognitiveservices-speech
boto3
elevenlabs
